Еще больше сломали проект (вынесли файлы моделей из проекта)
