+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,13 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Александр Поляков on 13.03.2023.
+//
+
+import Foundation
+private struct QuizQuestion {
+  let image: String
+  let text: String
+  let correctAnswer: Bool
+}
