+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Александр Поляков on 13.03.2023.
+//
+
+import Foundation
+// для состояния "Результат квиза"
+private struct QuizResultsViewModel {
+  let title: String
+  let text: String
+  let buttonText: String
+}
