+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Александр Поляков on 13.03.2023.
+//
+
+import UIKit
+
+// для состояния "Вопрос задан"
+private struct QuizStepViewModel {
+  let image: UIImage
+  let question: String
+  let questionNumber: String
+}
